requirements/app.txt
guniflask-cli>=0.12.14,<0.13
Flask-SQLAlchemy>=2.4.1
PyMySQL>=0.9.3
cryptography>=2.8
